memindahkan keamanan ke halaman Home dan mengosongkan  passphrase saat extensi dikunci dan dibuka
